# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for the EatMe recipe generation service",
        "title": "EatMe API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/thread/suggest": {
            "post": {
                "description": "Start a new suggestion thread",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "thread"
                ],
                "summary": "Start a new suggestion thread",
                "operationId": "startSuggestionThread",
                "parameters": [
                    {
                        "description": "Suggestion thread request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.StartSuggestionThreadRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ThreadState"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            }
        },
        "/thread/{threadId}": {
            "get": {
                "description": "Get a thread",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "thread"
                ],
                "summary": "Get a thread",
                "operationId": "getThread",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Thread ID",
                        "name": "threadId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ThreadState"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "404": {
                        "description": "Thread not found",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            }
        },
        "/thread/{threadId}/accept/{suggestionId}": {
            "post": {
                "description": "Accept a suggestion",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "thread"
                ],
                "summary": "Accept a suggestion",
                "operationId": "acceptSuggestion",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Thread ID",
                        "name": "threadId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Suggestion ID",
                        "name": "suggestionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserRecipe"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "404": {
                        "description": "Thread not found",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            }
        },
        "/thread/{threadId}/modify/chat": {
            "post": {
                "description": "Modify a recipe via chat",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "thread"
                ],
                "summary": "Modify a recipe via chat",
                "operationId": "modifyRecipe",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Thread ID",
                        "name": "threadId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Modify recipe via chat request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ModifyRecipeViaChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserRecipe"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "404": {
                        "description": "Thread not found",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            }
        },
        "/thread/{threadId}/question": {
            "post": {
                "description": "Answer a cooking question",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "thread"
                ],
                "summary": "Answer a cooking question",
                "operationId": "answerCookingQuestion",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Thread ID",
                        "name": "threadId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Answer cooking question request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AnswerCookingQuestionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AnswerCookingQuestionResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            }
        },
        "/thread/{threadId}/suggest": {
            "post": {
                "description": "Get new recipe suggestions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "thread"
                ],
                "summary": "Get new suggestions",
                "operationId": "getNewSuggestions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Thread ID",
                        "name": "threadId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Get new suggestions request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.GetNewSuggestionsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.RecipeSuggestion"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "404": {
                        "description": "Thread not found",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.APIError": {
            "description": "APIError represents an API error response",
            "type": "object",
            "required": [
                "code",
                "message"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.AnswerCookingQuestionRequest": {
            "description": "AnswerCookingQuestionRequest represents a request to answer a cooking question",
            "type": "object",
            "required": [
                "question"
            ],
            "properties": {
                "question": {
                    "type": "string"
                }
            }
        },
        "models.AnswerCookingQuestionResponse": {
            "description": "AnswerCookingQuestionResponse represents a response to an answer cooking question",
            "type": "object",
            "required": [
                "answer"
            ],
            "properties": {
                "answer": {
                    "type": "string"
                }
            }
        },
...
        "models.GetNewSuggestionsRequest": {
            "description": "GetNewSuggestionsRequest represents a request to get new suggestions",
            "type": "object",
            "properties": {
                "prompt": {
                    "type": "string"
                }
            }
        },
        "models.Ingredient": {
            "description": "Ingredient represents an ingredient in a recipe",
            "type": "object",
            "required": [
                "name",
                "quantity",
                "unit"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Flour"
                },
                "quantity": {
                    "type": "number",
                    "example": 1
                },
                "unit": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.MeasurementUnit"
                        }
                    ],
                    "example": "cup"
                }
            }
        },
        "models.MeasurementUnit": {
            "type": "string",
            "enum": [
                "g",
                "ml",
                "tsp",
                "tbsp",
                "cup",
                "oz",
                "lb"
            ],
            "x-enum-varnames": [
                "MeasurementUnitGram",
                "MeasurementUnitMilliliter",
                "MeasurementUnitTeaspoon",
                "MeasurementUnitTablespoon",
                "MeasurementUnitCup",
                "MeasurementUnitOunce",
                "MeasurementUnitPound"
            ]
        },
        "models.ModifyRecipeViaChatRequest": {
            "description": "ModifyRecipeViaChatRequest represents a request to modify a recipe via chat",
            "type": "object",
            "required": [
                "prompt"
            ],
            "properties": {
                "prompt": {
                    "type": "string"
                }
            }
        },
...
        "models.RecipeBody": {
            "description": "RecipeBody represents the contents of a recipe",
            "type": "object",
            "required": [
                "description",
                "ingredients",
                "servings",
                "steps",
                "title",
                "total_time_minutes"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "A classic Italian dish"
                },
                "ingredients": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Ingredient"
                    }
                },
                "servings": {
                    "type": "integer",
                    "example": 4
                },
                "steps": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string",
                    "example": "Veal Bolognese"
                },
                "total_time_minutes": {
                    "type": "integer",
                    "example": 120
                }
            }
        },
        "models.RecipeSuggestion": {
            "description": "RecipeSuggestion represents a suggestion for a recipe",
            "type": "object",
            "required": [
                "accepted",
                "created_at",
                "id",
                "response_text",
                "suggestion",
                "thread_id",
                "updated_at"
            ],
            "properties": {
                "accepted": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "response_text": {
                    "type": "string"
                },
                "suggestion": {
                    "$ref": "#/definitions/models.RecipeBody"
                },
                "thread_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.StartSuggestionThreadRequest": {
            "description": "StartSuggestionThreadRequest represents a request to start a suggestion thread",
            "type": "object",
            "required": [
                "prompt"
            ],
            "properties": {
                "prompt": {
                    "type": "string"
                }
            }
        },
        "models.ThreadState": {
            "description": "A thread of suggestions for a recipe",
            "type": "object",
            "required": [
                "created_at",
                "current_prompt",
                "id",
                "original_prompt",
                "suggestions",
                "updated_at"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "current_prompt": {
                    "type": "string"
                },
                "current_recipe": {
                    "$ref": "#/definitions/models.RecipeBody"
                },
                "id": {
                    "type": "string"
                },
                "original_prompt": {
                    "type": "string"
                },
                "recipe_id": {
                    "type": "string"
                },
                "suggestions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.RecipeSuggestion"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.UserRecipe": {
            "description": "UserRecipe is the user's personal copy (favorites, edits).",
            "type": "object",
            "required": [
                "created_at",
                "description",
                "id",
                "ingredients",
                "is_favorite",
                "latest_version_id",
                "servings",
                "steps",
                "thread_id",
                "title",
                "total_time_minutes",
                "updated_at",
                "user_id"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "example": "A classic Italian dish"
                },
                "global_recipe_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "ingredients": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Ingredient"
                    }
                },
                "is_favorite": {
                    "type": "boolean"
                },
                "latest_version_id": {
                    "type": "string"
                },
                "servings": {
                    "type": "integer",
                    "example": 4
                },
                "steps": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "thread_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "example": "Veal Bolognese"
                },
                "total_time_minutes": {
                    "type": "integer",
                    "example": 120
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        }
    }
}